+++ objects/media/entity.obj.c
@@ -184,7 +184,7 @@ d_define_method_override(entity, draw)(struct s_object *self, struct s_object *e
                           environment_attributes->camera_focus_x[environment_attributes->current_surface],
                           environment_attributes->camera_focus_y[environment_attributes->current_surface], environment_attributes->current_w,
                           environment_attributes->current_h, environment_attributes->zoom[environment_attributes->current_surface])))
-                while (((int)d_call(current_element->drawable, m_drawable_draw, environment)) == d_drawable_return_continue);
+                while (((intptr_t)d_call(current_element->drawable, m_drawable_draw, environment)) == d_drawable_return_continue);
             }
         }
       d_catch(exception)
