+++ objects/media/ui/uiable.obj.c
@@ -227,7 +227,7 @@ d_define_method_override(uiable, draw)(struct s_object *self, struct s_object *e
       /* doesn't inerith the flip (this object, the uiable, doesn't flip) and the zoom is hardcoded to one */
       d_call(uiable_attributes->background[uiable_attributes->selected_mode][index], m_drawable_keep_scale, environment_attributes->current_w,
              environment_attributes->current_h);
-      while (((int) d_call(uiable_attributes->background[uiable_attributes->selected_mode][index], m_drawable_draw, environment)) ==
+      while (((intptr_t)d_call(uiable_attributes->background[uiable_attributes->selected_mode][index], m_drawable_draw, environment)) ==
              d_drawable_return_continue);
     }
   if (uiable_attributes->background_mask_A) { /* only if visible */
